receive_news/requirements.txt
anyio==3.6.1
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
elastic-transport==8.4.0
elasticsearch==8.4.2
expertai-nlapi==2.4.1
fastapi==0.85.0
h11==0.14.0
idna==3.4
kafka-python==2.0.2
oauthlib==3.2.1
pydantic==1.10.2
python-dotenv==0.21.0
requests==2.28.1
requests-oauthlib==1.3.1
sniffio==1.3.0
starlette==0.20.4
tweepy==4.10.1
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
